+++ src/components/Input/index.tsx
@@ -4,4 +4,3 @@ export { FileInput } from './FileInput'
 export { Select } from './Select'
 export { InputWithBadge } from './InputWithBadge'
 export { DateInput } from './DateInput'
-export { Textarea } from './Textarea'
