+++ src/components/JournalTable/JournalTable.tsx
@@ -1,9 +1,10 @@
 import React, { RefObject, useContext, useMemo, useState } from 'react'
+import { JournalContext } from '../../contexts/JournalContext'
 import DownloadCloud from '../../icons/DownloadCloud'
 import { Button, ButtonVariant } from '../Button'
 import { Header } from '../Container'
 import { DataState, DataStateStatus } from '../DataState'
-import { JournalContext, View } from '../Journal'
+import { View } from '../Journal'
 import { JournalConfig } from '../Journal/Journal'
 import { JournalRow } from '../JournalRow'
 import { JournalSidebar } from '../JournalSidebar'
