--- src/components/Textarea/index.tsx
@@ -1,1 +1,0 @@
-export { Textarea } from './Textarea'
