+++ src/contexts/JournalContext/JournalContext.tsx
@@ -1,0 +1,24 @@
+import { createContext } from 'react'
+import { useJournal } from '../../hooks/useJournal'
+
+export type JournalContextType = ReturnType<typeof useJournal>
+export const JournalContext = createContext<JournalContextType>({
+  data: undefined,
+  isLoading: false,
+  error: undefined,
+  refetch: () => {},
+  selectedEntryId: undefined,
+  setSelectedEntryId: () => {},
+  closeSelectedEntry: () => {},
+  addEntry: () => {},
+  addEntryLine: () => {},
+  removeEntryLine: () => {},
+  create: () => {},
+  changeFormData: () => {},
+  submitForm: () => {},
+  cancelForm: () => {},
+  form: undefined,
+  setForm: () => {},
+  sendingForm: false,
+  apiError: undefined,
+})
