+++ src/contexts/JournalContext/index.tsx
@@ -1,0 +1,1 @@
+export { JournalContext } from './JournalContext'
